+++ inc/BitFunnel/TermInfo.h
@@ -4,7 +4,7 @@
 #include "BitFunnel/ITermTable.h"         // Embeds ITermTable::TermKind.
 #include "BitFunnel/Index/IFactSet.h"           // FactHandle is used as a parameter.
 #include "BitFunnel/NonCopyable.h"        // TermInfo inherits from NonCopyable.
-#include "Term.h"               // Embeds Term::Hash.
+#include "BitFunnel/Term.h"               // Embeds Term::Hash.
 
 
 namespace BitFunnel
@@ -54,16 +54,16 @@ namespace BitFunnel
         const ITermTable& m_termTable;
 
         // Slot index for first RowId.
-        unsigned m_rowIdStart;
+        size_t m_rowIdStart;
 
         // Number of RowId to be enumerated.
-        unsigned m_rowIdCount;
+        size_t m_rowIdCount;
 
         // Stores what kind of the term this TermInfo is for.
         ITermTable::TermKind m_termKind;
 
         // Using int, rather than unsigned to allow for -1 sentinal for
         // position before start of enumerator.
-        int m_currentRow;
+        int64_t m_currentRow;
     };
 }
